pyproject.toml
[build-system]
requires = ["hatchling", "pyyaml"]
build-backend = "hatchling.build"

[project]
name = "ymery"
version = "0.0.62"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "httpx>=0.28.1",
    "imgui-bundle>=1.92.4",
    "matplotlib>=3.8.0",
    "munch>=4.0.0",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
audio = [
    "soundfile>=0.13.1",
]
dev = [
    "pytest>=9.0.2",
    "twine>=6.2.0",
]

[project.scripts]
ymery = "ymery.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ymery"]

[tool.hatch.build.targets.wheel.hooks.custom]
path = "scripts/hatch_build.py"

[tool.hatch.build.targets.wheel.force-include]
"demo" = "ymery/demo"
"demo-aggregated" = "ymery/demo_aggregated"

[tool.hatch.build.targets.sdist]
include = [
    "src/ymery/**/*.py",
    "src/ymery/**/*.yaml",
    "demo/**/*",
    "scripts/hatch_build.py",
]

[dependency-groups]
dev = [
    "twine>=6.2.0",
]
